% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\alias{tui_theme}
\alias{set_tui_theme}
\title{Use a custom theme or set global theme}
\usage{
tui_theme(tui, ..., name = NULL)

set_tui_theme(name, ...)
}
\arguments{
\item{tui}{A \code{tuichart} \code{htmlwidget} object.}

\item{...}{Parameters to define theme, see \url{http://nhn.github.io/tui.chart/latest/tui.chart#registerTheme}.}

\item{name}{The name of the theme.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Use a custom theme or set global theme
}
