% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tui-options}
\alias{tui-options}
\alias{tui_legend}
\alias{tui_yAxis}
\alias{tui_xAxis}
\alias{tui_chart}
\alias{tui_series}
\alias{tui_tooltip}
\alias{tui_plot}
\alias{tui_chartExportMenu}
\title{Set options for a chart}
\usage{
tui_legend(tui, ...)

tui_yAxis(tui, ...)

tui_xAxis(tui, ...)

tui_chart(tui, ...)

tui_series(tui, ...)

tui_tooltip(tui, ...)

tui_plot(tui, ...)

tui_chartExportMenu(tui, ...)
}
\arguments{
\item{tui}{A \code{tuichart} \code{htmlwidget} object.}

\item{...}{Named parameters, see \url{http://nhn.github.io/tui.chart/latest/tui.chart}.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Set options for a chart
}
