% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuichart.R
\name{tuichart}
\alias{tuichart}
\title{Create a \code{tuichart} htmlwidget}
\usage{
tuichart(type = "bar", data = NULL, options = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{type}{Type of chart.}

\item{data}{A \code{list} of parameters for the data used in the chart.}

\item{options}{A \code{list} of options for the chart.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Create a \code{tuichart} htmlwidget
}
