% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tuichartr-package}
\alias{tuichartr-package}
\title{An \code{htmlwidget} interface to the
tui-chart javascript chart library}
\description{
This package allow you to use tui-chart.js (\url{https://ui.toast.com/tui-chart}),
to create beautiful chart for data visualization.
}
\author{
Victor Perrier (@dreamRs_fr)
}
