% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuimaps.R
\name{tuimap-shiny}
\alias{tuimap-shiny}
\alias{tuimapOutput}
\alias{renderTuimap}
\title{Shiny bindings for tuimap}
\usage{
tuimapOutput(outputId, width = "100\%", height = "400px")

renderTuimap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a tuimaps}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using tuimaps within Shiny
applications and interactive Rmd documents.
}
\examples{
if (interactive()) {

  library(shiny)
  library(rnaturalearth)
  library(dplyr)
  library(tuichartr)

  # Retrieve world map
  world <- ne_countries(returnclass = "sf") \%>\%
    filter(continent != "Antarctica")

  ui <- fluidPage(
    tags$h2("Include tuimap in Shiny"),
    fluidRow(
      column(
        width = 3,
        actionButton(
          inputId = "refresh",
          label = "Refresh data"
        )
      ),
      column(
        width = 9,
        tuimapOutput(outputId = "my_map")
      )
    )
  )

  server <- function(input, output, session) {

    output$my_map <- renderTuimap({

      input$refresh

      # add a random numeric variable
      world$random <- sample(1:100, nrow(world), TRUE)

      # draw map
      tuimap() \%>\%
        add_map_data(
          data = world,
          mapping = aes(code = adm0_a3, label = name, value = random)
        ) \%>\%
        tui_chart(title = "World map (minus Antarctica)")

    })

  }

  shinyApp(ui, server)
}
}
