% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuimaps.R
\name{tuimaps}
\alias{tuimaps}
\title{Create a \code{tuichart} htmlwidget}
\usage{
tuimaps(shape, code, name, data = NULL, options = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{shape}{An \code{sf} object.}

\item{code}{Variable to use as unique identifier for the polygons.}

\item{name}{Varaible to use as label for the polygons.}

\item{data}{A \code{list} of parameters for the data used in the chart.}

\item{options}{A \code{list} of options for the chart.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Create a \code{tuichart} htmlwidget
}
