% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data.R
\name{add-data}
\alias{add-data}
\alias{add_data}
\alias{add_default_data}
\alias{add_pie_data}
\alias{add_line_data}
\alias{add_scatter_data}
\alias{add_heat_data}
\alias{add_tree_data}
\alias{add_boxplot_data}
\alias{add_map_data}
\title{Add data to a \code{tuichart}}
\usage{
add_data(tui, data, mapping)

add_default_data(tui, data, mapping)

add_pie_data(tui, data, mapping)

add_line_data(tui, data, mapping)

add_scatter_data(tui, data, mapping)

add_heat_data(tui, data, mapping)

add_tree_data(tui, data, mapping)

add_boxplot_data(tui, data, mapping)

add_map_data(tui, data, mapping)
}
\arguments{
\item{tui}{A \code{tuichart} \code{htmlwidget} object.}

\item{data}{A \code{data.frame}, or an object coercible to \code{data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Add data to a \code{tuichart}
}
\details{
Aesthetics vary according to the type of graph:
 \itemize{
  \item \strong{bar, column, line, area, radial, boxplot} : x, y, and group (optional).
  \item \strong{pie} : x, y.
  \item \strong{heatmap} : x, y, value.
  \item \strong{treemap} : leval1, (level2, level3, ...), value
  \item \strong{scatter, bubble} : x, y, group (optional), label (optional), size (bubble).
 }
}
