% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-data.R
\name{add-data}
\alias{add-data}
\alias{add_data}
\alias{add_scatter_data}
\alias{add_heat_data}
\title{Add data to a \code{tuichart}}
\usage{
add_data(tui, data, mapping)

add_scatter_data(tui, data, mapping)

add_heat_data(tui, data, mapping)
}
\arguments{
\item{tui}{A \code{tuichart} \code{htmlwidget} object.}

\item{data}{A \code{data.frame}, or an object coercible to \code{data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Add data to a \code{tuichart}
}
