% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuichart.R
\name{tuichart}
\alias{tuichart}
\title{Create a \code{tuichart} htmlwidget}
\usage{
tuichart(type = "bar", data = NULL, options = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{type}{Type of chart.}

\item{data}{A \code{list} of parameters for the data used in the chart.}

\item{options}{A \code{list} of options for the chart.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
A \code{tuichart} \code{htmlwidget} object.
}
\description{
Create a \code{tuichart} htmlwidget
}
\examples{

library(gapminder)
library(dplyr)

n_countries <- gapminder_unfiltered \%>\%
  filter(year == 2007) \%>\%
  count(continent)

# a barchart
tuichart("column") \%>\%
  add_data(n_countries, aes(x = continent, y = n)) \%>\%
  tui_series(showLabel = TRUE)

# a treemap
tuichart("treemap") \%>\%
  add_data(n_countries, aes(level1 = continent, value = n)) \%>\%
  tui_series(showLabel = TRUE)

# a pie chart
tuichart("pie") \%>\%
  add_data(n_countries, aes(x = continent, y = n)) \%>\%
  tui_series(showLabel = TRUE)



gdp_italy <- gapminder_unfiltered \%>\%
  filter(country == "Italy")

# line chart
tuichart("line") \%>\%
  add_data(gdp_italy, aes(x = year, y = gdpPercap)) \%>\%
  tui_chart(
    title = "A chart",
    format = "0.00"
  ) \%>\%
  tui_series(
    zoomable = TRUE
  ) \%>\%
  tui_plot(
    hideLine = TRUE
  ) \%>\%
  tui_xAxis(
    tickInterval = "auto",
    title = "X axis title"
  )

}
